Change generics to BigInteger.

+++ src/main/kotlin/harpo/network/communication/p2p/infrastructure/grpc/repository/OperationsRepository.kt
@@ -9,11 +9,12 @@ import harpo.network.communication.p2p.infrastructure.grpc.client.ChannelPool
 import io.ep2p.kademlia.node.KademliaNodeAPI
 import io.ep2p.kademlia.node.Node
 import java.math.BigDecimal
+import java.math.BigInteger
 import harpo.network.communication.p2p.Node as NodeGRPC
 
 class OperationsRepository {
 
-    fun ping(receiver: Node<BigDecimal, ConnectionInfoImpl>?, self: KademliaNodeAPI<BigDecimal, ConnectionInfoImpl>?): Answer {
+    fun ping(receiver: Node<BigInteger, ConnectionInfoImpl>?, self: KademliaNodeAPI<BigInteger, ConnectionInfoImpl>?): Answer {
         val channel = ChannelPool.getBy(Contact(receiver?.connectionInfo?.ip!!, receiver.connectionInfo.port))
         val stub = KademliaServiceGrpc.newBlockingStub(channel)
         val contactGRPC = newBuilder().setIp(self?.connectionInfo?.ip).setPort(self?.connectionInfo?.port!!).build()
@@ -25,7 +26,7 @@ class OperationsRepository {
         return stub.ping(grpcSelfNode)
     }
 
-    fun shutdown(receiver: Node<BigDecimal, ConnectionInfoImpl>?, self: KademliaNodeAPI<BigDecimal, ConnectionInfoImpl>?) {
+    fun shutdown(receiver: Node<BigInteger, ConnectionInfoImpl>?, self: KademliaNodeAPI<BigInteger, ConnectionInfoImpl>?) {
         val channel = ChannelPool.getBy(Contact(receiver?.connectionInfo?.ip!!, receiver.connectionInfo.port))
         val stub = KademliaServiceGrpc.newBlockingStub(channel)
         val contactGRPC = newBuilder().setIp(self?.connectionInfo?.ip).setPort(self?.connectionInfo?.port!!).build()
@@ -37,7 +38,7 @@ class OperationsRepository {
         stub.shutdownSignal(grpcSelfNode)
     }
 
-    fun closest(receiver: Node<BigDecimal, ConnectionInfoImpl>?, self: KademliaNodeAPI<BigDecimal, ConnectionInfoImpl>?): List<NodeGRPC> {
+    fun closest(receiver: Node<BigInteger, ConnectionInfoImpl>?, self: KademliaNodeAPI<BigInteger, ConnectionInfoImpl>?): List<NodeGRPC> {
         val channel = ChannelPool.getBy(Contact(receiver?.connectionInfo?.ip!!, receiver.connectionInfo.port))
         val stub = KademliaServiceGrpc.newBlockingStub(channel)
         val contactGRPC = newBuilder().setIp(self?.connectionInfo?.ip).setPort(self?.connectionInfo?.port!!).build()
