+++ src/main/kotlin/harpo/network/communication/p2p/infrastructure/messages/type/Message.kt
@@ -5,15 +5,15 @@ import io.ep2p.kademlia.node.KademliaNodeAPI
 import io.ep2p.kademlia.node.Node
 import io.ep2p.kademlia.protocol.message.KademliaMessage
 import java.io.Serializable
-import java.math.BigDecimal
+import java.math.BigInteger
 
 interface Message {
 
     fun <INPUT : Serializable?, OUTPUT : Serializable?> sendMessage(
-        api: KademliaNodeAPI<BigDecimal, ConnectionInfoImpl>?,
-        external: Node<BigDecimal, ConnectionInfoImpl>?,
-        message: KademliaMessage<BigDecimal, ConnectionInfoImpl, OUTPUT>?
-    ): KademliaMessage<BigDecimal, ConnectionInfoImpl, INPUT>
+        api: KademliaNodeAPI<BigInteger, ConnectionInfoImpl>?,
+        external: Node<BigInteger, ConnectionInfoImpl>?,
+        message: KademliaMessage<BigInteger, ConnectionInfoImpl, OUTPUT>?
+    ): KademliaMessage<BigInteger, ConnectionInfoImpl, INPUT>
 
 }
